use workerpool for scaling images
